Moved to ...integration.containerutils
